Rework Templates
This re-work of the library's templates has two objectives
1. Permit custom allocators with input pointclouds and output features
using that same allocation. Required for working with PCL.
2. Reorder template arguments. PointType and Alloc are always deducable
from function arguments. Accessor is however, not, it is therefore the
only one a user will need to specify. Hence, we order it first so the
user never has to specify PointType or Alloc